requirements.txt
fastapi==0.136.0
uvicorn==0.34.0
httpx==0.28.1
pydantic==2.10.6
python-dotenv==1.2.2
